Backend/core/requirements.txt
asgiref==3.6.0
audioread==3.0.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
emoji==2.2.0
filelock==3.10.7
fonttools==4.39.0
huggingface-hub==0.0.12
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
install==1.3.5
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.1
librosa==0.9.2
llvmlite==0.39.1
matplotlib==3.6.3
mongoengine==0.27.0
msgpack==1.0.4
numba==0.56.4
numpy==1.19.5
packaging==23.0
pandas==1.4.4
Pillow==9.4.0
platformdirs==3.1.0
pooch==1.7.0
pycparser==2.21
pydub==0.25.1
pymongo==4.3.3
pyparsing==3.0.9
pythainlp==3.1.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
resampy==0.4.2
sacremoses==0.0.53
scikit-learn==1.2.1
scipy==1.10.0
six==1.16.0
sklearn==0.0.post1
soundfile==0.12.1
soxr==0.3.4
SpeechRecognition==3.9.0
sqlparse==0.2.4
thai-sentiment==0.1.3
threadpoolctl==3.1.0
tokenizers==0.10.3
torch==1.13.1
torchaudio==0.13.1
tqdm==4.65.0
transformers==4.9.1
typing_extensions==4.5.0
urllib3==1.26.14
whitenoise==6.4.0
zipp==3.15.0
